Updated table,graph changes on scratch.py using lf_graph & lf_report libraries and file renamed as wifi_diag.py
